Add files via upload
This a BLIND BID .

Here is the link if you don't know what is BLIND BID ( https://en.wikipedia.org/wiki/First-price_sealed-bid_auction ).

This code is mixture of:
- DICTIONARY
- DEF function
- FOR loop
- IF-ELIF condition
- WHILE loop
- OS module

In simple terms, BID is nothing but AUCTION. The person who gives highest price to purchase thing will be his/her. BLIND BID is where we know that there are people to bid but we will NOT GET KNOW TO THAT WHICH PERSON HAS BID WAHT PRICE.
Let's say we have 3 person. First it'll ask name and bid to first person, then system will ask is there any other bidder? we have type yes as we have. it'll clear bid of first person and the steps repeats. At last when we type no it'll check the highest bid and will print the name with bid.

IF POSSIBLE SEND FEEDBACK FOR IMPROVEMENT.